Remove private methods from public __init__
